Add files via upload
A bunch of Python codes which I find useful on the web. Great to use for fast analysis of time series data.

In all, great for aggregating daily data into monthly data. Then use the data to do Time Series Analysis with ARIMA and then find the optimal model to use for subsequent forecasts.